Python files
These are the python-written files used for the project. "funtionsvd.py" takes care of the image processing and the singular value decomposition logic behind the compression of the images, while "interfaceGUI.py" works for running the project with an interface in which the user may upload the images they want to compress.
Further explanation of the theory behind this project is within the project report file.